requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.4.4
dnspython==2.6.1
exceptiongroup==1.2.0
fastapi==0.110.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
Mako==1.3.3
MarkupSafe==2.1.5
mongomock==4.1.2
packaging==24.0
parse==1.20.1
parse-type==0.6.2
pluggy==1.4.0
pydantic==2.7.0
pydantic-core==2.18.1
pymongo==4.6.3
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-bdd==7.1.2
pytest-cov==5.0.0
pytest-mock==3.14.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
sentinels==1.0.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tomli==2.0.1
typing-extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
